requirements.txt
pyautogen
openai==0.28.1
python-dotenv
requests
beautifulsoup4
selenium
webdriver_manager
langchain